chore: manage init function per provider
Per comment https://github.com/redhat-developer/mapt/pull/435#discussion_r2005170252 we may missing a required initialization on azure provider, to ensure all initializations happens on each provider we added an interface which each provider should implement and will contain all required initializacion specific to the provider. That initialization function is called from context initalization.
