+++ pkg/manager/context/context.go
@@ -2,12 +2,10 @@ package context
 
 import (
 	"fmt"
-	"os"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 	"github.com/redhat-developer/mapt/pkg/integrations/cirrus"
 	"github.com/redhat-developer/mapt/pkg/integrations/github"
-	"github.com/redhat-developer/mapt/pkg/provider/aws/data"
 	"github.com/redhat-developer/mapt/pkg/util"
 	"github.com/redhat-developer/mapt/pkg/util/logging"
 	utilMaps "github.com/redhat-developer/mapt/pkg/util/maps"
@@ -61,7 +59,11 @@ type context struct {
 // mapt context
 var mc *context
 
-func Init(ca *ContextArgs) error {
+type Provider interface {
+	Init(backedURL string) error
+}
+
+func Init(ca *ContextArgs, provider Provider) error {
 	mc = &context{
 		runID:         CreateRunID(),
 		projectName:   ca.ProjectName,
@@ -73,8 +75,8 @@ func Init(ca *ContextArgs) error {
 		serverless:    ca.Serverless,
 	}
 	addCommonTags()
-	// Manage remote state requirements
-	if err := manageRemoteState(ca.BackedURL); err != nil {
+	// Init provider
+	if err := provider.Init(ca.BackedURL); err != nil {
 		return err
 	}
 	// Manage integrations
@@ -142,28 +144,6 @@ func addCommonTags() {
 	}
 	mc.tags[tagKeyOrigin] = origin
 	mc.tags[TagKeyProjectName] = mc.projectName
-}
-
-// Under some circumstances it is poosible we need to update Location initial configuration
-// due to usage of remote backed url. i.e. https://github.com/redhat-developer/mapt/issues/392
-
-// This function will check if backed url is remote and if so change initial values to be able to
-// use it.
-func manageRemoteState(backedURL string) error {
-	if data.ValidateS3Path(backedURL) {
-		awsRegion, err := data.GetBucketLocationFromS3Path(backedURL)
-		if err != nil {
-			return err
-		}
-		if err := os.Setenv("AWS_DEFAULT_REGION", *awsRegion); err != nil {
-			return err
-		}
-		if err := os.Setenv("AWS_REGION", *awsRegion); err != nil {
-			return err
-		}
-		return nil
-	}
-	return nil
 }
 
 func manageIntegration(ca *ContextArgs) error {
