+++ pkg/provider/azure/action/windows/windows.go
@@ -58,7 +58,7 @@ type ghActionsRunnerData struct {
 
 func Create(ctx *maptContext.ContextArgs, r *WindowsRequest) (err error) {
 	// Create mapt Context
-	if err := maptContext.Init(ctx); err != nil {
+	if err := maptContext.Init(ctx, azure.Provider()); err != nil {
 		return err
 	}
 
@@ -84,7 +84,7 @@ func Create(ctx *maptContext.ContextArgs, r *WindowsRequest) (err error) {
 
 func Destroy(ctx *maptContext.ContextArgs) error {
 	// Create mapt Context
-	if err := maptContext.Init(ctx); err != nil {
+	if err := maptContext.Init(ctx, azure.Provider()); err != nil {
 		return err
 	}
 	// destroy
